32 Move function to context.
