+++ src/App.js
@@ -1,9 +1,9 @@
-import {v4 as uuidv4} from 'uuid'
+
 import {BrowserRouter as Router, Route, Routes} from 'react-router-dom'
 import FeedbackList from "./components/FeedbackList"
 import Header from "./components/Header"
-import { useState } from "react"
-import FeedbackData from './data/FeedbackData'
+//import { useState } from "react"
+//import FeedbackData from './data/FeedbackData'
 import FeedbackStats from "./components/FeedbackStats"
 import FeedbackForm from "./components/FeedbackForm"
 import AboutPage from './pages/AboutPage'
@@ -13,21 +13,11 @@ import { FeedbackProvider } from './context/FeedbackContext'
 
 function App(){
 
-    const[feedbackState, setFeedbackState] = useState(FeedbackData)
+    //const [feedback, setFeedback]  = useState(FeedbackData)
 
-    const deleteFeedback = (id)=> {
-        if(window.confirm("Are you sure?")){
-            setFeedbackState(feedbackState.filter((item)=> item.id!==id))
-        }
-    }
+    
 
-    const addFeedback = (newFeedback) =>{
-        newFeedback.id = uuidv4()
-        console.log('App.js')
-        console.log(newFeedback)
-        setFeedbackState([newFeedback,...feedbackState])
-
-    }
+    
 return(
     <FeedbackProvider>
     <Router>
@@ -36,9 +26,9 @@ return(
             <Routes>
             <Route exact path='/' element={
                 <>
-                    <FeedbackForm handleAdd={addFeedback}/>
+                    <FeedbackForm /*handleAdd={addFeedback}*//>
                     <FeedbackStats /*feedback={feedbackState}*//>
-                    <FeedbackList /*feedback={feedbackState}*/ handleDelete={deleteFeedback}/>
+                    <FeedbackList /*feedback={feedbackState} handleDelete={deleteFeedback}*//>
                 </>
             }>
