+++ src/components/FeedbackForm.jsx
@@ -2,13 +2,17 @@ import { useState } from "react"
 import Card from "./shared/Card"
 import Button from "./shared/Button"
 import RatingSelect from "./RatingSelect"
+import { useContext } from 'react'
+import FeedbackContext from '../context/FeedbackContext'
 
-function FeedbackForm({handleAdd}){
+function FeedbackForm(/*{handleAdd}*/){
 
     const[text, setText] = useState('')
     const[btnDisbaled, setBtnDisbaled] = useState(true)
     const[message, setMessage] = useState('')
     const[rating, setRating] = useState(10)
+
+    const{addFeedback} = useContext(FeedbackContext)
     
 
     const handleTextChange = (e) => {
@@ -33,7 +37,8 @@ function FeedbackForm({handleAdd}){
                 text, rating
             }
 
-            handleAdd(newFeedback)
+            //handleAdd(newFeedback)
+            addFeedback(newFeedback)
 
             setText('')
         }
