+++ src/context/FeedbackContext.js
@@ -1,4 +1,5 @@
 import { createContext, useState } from "react"
+import {v4 as uuidv4} from 'uuid'
 
 const FeedbackContext = createContext()
 
@@ -10,14 +11,39 @@ export const FeedbackProvider = ({children}) => {
         [
             {
                 id : 1,
-                text : 'this is coming from the context',
+                text : 'This is feedback item 1',
                 rating : 10
+            },
+            {
+                id : 2,
+                text : 'This is feedback item 2',
+                rating : 8
+            },
+            {
+                id : 3,
+                text : 'This is feedback item 3',
+                rating : 5
             }
         ]
     )
-    
+
+    const deleteFeedback = (id)=> {
+        console.log('Hello')
+        if(window.confirm("Are you sure?")){
+            setFeedback(feedback.filter((item)=> item.id!==id))
+        }
+    }
+
+    const addFeedback = (newFeedback) =>{
+        newFeedback.id = uuidv4()
+        console.log('App.js')
+        console.log(newFeedback)
+        setFeedback([newFeedback,...feedback])
+
+    }
+
     return <FeedbackContext.Provider value= {{
-        feedback,
+        feedback, deleteFeedback, addFeedback
     }}>
         {children}
     </FeedbackContext.Provider>
